requirements.txt
icalendar==6.3.1
Jinja2==3.1.6
pypandoc==1.15
python-docx==1.2.0
PyYAML==6.0.2
